dev: change feature

+++ .eslintrc.js
@@ -202,7 +202,7 @@ module.exports = {
     'padded-blocks': [2, 'never'],
     quotes: [
       2,
-      'single',
+      'double',
       {
         avoidEscape: true,
         allowTemplateLiterals: true
